Refactor the delete blog operation so that a blog can be deleted only by the user who added it. Therefore, deleting a blog is possible only if the token sent with the request is the same as that of the blog's creator in part 4

+++ part4/controllers/blogs.js
@@ -19,7 +19,7 @@ blogsRouter.post('/', async (request, response, next) => {
   try {
     const body = request.body
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    if(!decodedToken.id) {
+    if (!decodedToken.id) {
       return response.status(401).json({ error: 'token invalid' })
     }
     const user = await User.findById(decodedToken.id)
@@ -49,8 +49,24 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+
+    if (!user) {
+      return response.status(400).json({ error: 'userId missing or invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (user._id.toString() === blog.user.toString()) {
+      await Blog.findByIdAndDelete(request.params.id)
+      return response.status(204).end()
+    }
+
+    return response.status(401).json({ error: 'cannot delete other users blogs' })
   } catch (exception) {
     next(exception)
   }
